requirements.txt
aiofiles==0.7.0
alembic==1.7.1
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dynaconf==3.1.5
ecdsa==0.17.0
fastapi==0.68.1
flake8==3.9.2
flake8-bugbear==21.4.3
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-variables-names==0.0.4
greenlet==1.1.1
h11==0.12.0
idna==3.2
importlib-resources==5.2.2
iniconfig==1.1.1
itsdangerous==2.0.1
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
passlib==1.7.4
pluggy==1.0.0
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pydantic==1.8.2
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.5
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
wincertstore==0.2
zipp==3.5.0
